+++ CHANGELOG.md
@@ -39,6 +39,7 @@
   showing or extracting text
 * Incorrect handling of page rotation values in `hexapdf merge`
 * Missing handling of certain rotation values in `hexapdf modify`
+* Removal of unused pages in `hexapdf modify`
 
 
 ## 0.3.0 - 2017-01-25